WIP Moved highlighted notes state to app level
