+++ src/components/Guitar/Guitar.svelte
@@ -1,29 +1,29 @@
 <script lang="ts">
-  import { Config, HighlightConfig, String } from './subcomponents';
+  import { Config, String } from './subcomponents';
   import type { SelectedNote } from './types';
-  import { assertEventTargetIsNode } from '@/utils';
+  // import { assertEventTargetIsNode } from '@/utils';
 
   export let scale: string[] = [];
+  export let selectedNote: SelectedNote | null = null;
+  export let highlightedNotes: SelectedNote[] = [];
 
   let numberOfFrets = 22;
   let numberOfStrings = 6;
   let stringTunings = ['E', 'A', 'D', 'G', 'B', 'E'];
   let displayConfig = false;
 
-  let selectedNote: SelectedNote | null = null;
-  let highlightedNotes: SelectedNote[] = [];
 
   let guitarElement: HTMLElement;
 
-  function handleWindowClick({ target }: MouseEvent) {
-    assertEventTargetIsNode(target);
-    if (!guitarElement.contains(target)) {
-      selectedNote = null;
-    }
-  }
+  // function handleWindowClick({ target }: MouseEvent) {
+  //   assertEventTargetIsNode(target);
+  //   if (!guitarElement.contains(target)) {
+  //     selectedNote = null;
+  //   }
+  // }
 </script>
 
-<svelte:window on:click={handleWindowClick} />
+<!-- <svelte:window on:click={handleWindowClick} /> -->
 
 <div bind:this={guitarElement}>
   <div class="guitar">
@@ -44,12 +44,6 @@
       />
     {/each}
   </div>
-  {#if selectedNote}
-    <HighlightConfig
-      {selectedNote}
-      bind:highlightedNotes={highlightedNotes}
-    />
-  {/if}
   <button
     on:click={() => displayConfig = !displayConfig}
     type="button"
